Renamed folders to use underscores
Also modified main.cpp to reflect these changes.
